organised helper functions into dirs in decision tree dir.
