+++ src/mlpack/methods/decision_tree/select_functions/random_dimension_select.hpp
@@ -1,39 +1,35 @@
 /**
- * @file methods/decision_tree/all_dimension_select.hpp
+ * @file methods/decision_tree/select_functions/random_dimension_select.hpp
  * @author Ryan Curtin
  *
- * Selects all dimensions for a split.
+ * Selects one single random dimension to split on.
  *
  * mlpack is free software; you may redistribute it and/or modify it under the
  * terms of the 3-clause BSD license.  You should have received a copy of the
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef MLPACK_METHODS_DECISION_TREE_ALL_DIMENSION_SELECT_HPP
-#define MLPACK_METHODS_DECISION_TREE_ALL_DIMENSION_SELECT_HPP
+#ifndef MLPACK_METHODS_DECISION_TREE_RANDOM_DIMENSION_SELECT_HPP
+#define MLPACK_METHODS_DECISION_TREE_RANDOM_DIMENSION_SELECT_HPP
 
 namespace mlpack {
 
 /**
- * This dimension selection policy allows any dimension to be selected for
- * splitting.
+ * This dimension selection policy only selects one single random dimension.
  */
-class AllDimensionSelect
+class RandomDimensionSelect
 {
  public:
   /**
-   * Construct the AllDimensionSelect object for the given number of dimensions.
+   * Construct the RandomDimensionSelect object with the given number of
+   * dimensions.
    */
-  AllDimensionSelect() : i(0), dimensions(0) { }
+  RandomDimensionSelect() : dimensions(0) { }
 
   /**
    * Get the first dimension to select from.
    */
-  size_t Begin()
-  {
-    i = 0;
-    return 0;
-  }
+  size_t Begin() const { return RandInt(dimensions); }
 
   /**
    * Get the last dimension to select from.
@@ -41,18 +37,17 @@ class AllDimensionSelect
   size_t End() const { return dimensions; }
 
   /**
-   * Get the next dimension.
+   * Get the next (last) dimensions.  We only allow one dimension, so any 'next'
+   * dimension is past our bounds.
    */
-  size_t Next() { return ++i; }
+  size_t Next() const { return dimensions; }
 
   //! Get the number of dimensions.
   size_t Dimensions() const { return dimensions; }
-  //! Modify the number of dimensions.
+  //! Set the number of dimensions.
   size_t& Dimensions() { return dimensions; }
 
  private:
-  //! The current dimension we are looking at.
-  size_t i;
   //! The number of dimensions to select from.
   size_t dimensions;
 };
