+++ src/mlpack/methods/decision_tree/split_functions/best_binary_numeric_split_impl.hpp
@@ -1,5 +1,5 @@
 /**
- * @file methods/decision_tree/best_binary_numeric_split_impl.hpp
+ * @file methods/decision_tree/split_functions/best_binary_numeric_split_impl.hpp
  * @author Ryan Curtin
  *
  * Implementation of strategy that finds the best binary numeric split.
