+++ src/mlpack/methods/decision_tree/split_functions/random_binary_numeric_split.hpp
@@ -1,32 +1,30 @@
 /**
- * @file methods/decision_tree/all_categorical_split.hpp
- * @author Ryan Curtin
+ * @file methods/decision_tree/split_functions/random_binary_numeric_split.hpp
+ * @author Rishabh Garg
  *
- * This file defines a tree splitter that split a categorical feature into all
- * of the possible categories.
+ * A tree splitter that finds a random binary numeric split.
  *
  * mlpack is free software; you may redistribute it and/or modify it under the
  * terms of the 3-clause BSD license.  You should have received a copy of the
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef MLPACK_METHODS_DECISION_TREE_ALL_CATEGORICAL_SPLIT_HPP
-#define MLPACK_METHODS_DECISION_TREE_ALL_CATEGORICAL_SPLIT_HPP
+#ifndef MLPACK_METHODS_DECISION_TREE_RANDOM_BINARY_NUMERIC_SPLIT_HPP
+#define MLPACK_METHODS_DECISION_TREE_RANDOM_BINARY_NUMERIC_SPLIT_HPP
 
 #include <mlpack/prereqs.hpp>
 
 namespace mlpack {
 
 /**
- * The AllCategoricalSplit is a splitting function that will split categorical
- * features into many children: one child for each category. This is a generic
- * splitting strategy and can be used for both regression and classification
- * trees.
+ * The RandomBinaryNumericSplit is a splitting function for decision trees that
+ * will split based on a randomly selected point between the minimum
+ * and maximum value of the numerical dimension.
  *
- * @tparam FitnessFunction Fitness function to evaluate gain with.
+ * @tparam FitnessFunction Fitness function to use to calculate gain.
  */
 template<typename FitnessFunction>
-class AllCategoricalSplit
+class RandomBinaryNumericSplit
 {
  public:
   // No extra info needed for split.
@@ -35,91 +33,115 @@ class AllCategoricalSplit
   /**
    * Check if we can split a node.  If we can split a node in a way that
    * improves on 'bestGain', then we return the improved gain.  Otherwise we
-   * return the value 'bestGain'.  If a split is made, then splitInfo and
-   * aux may be modified.  For this particular split type, aux will be empty
-   * and splitInfo will store the number of children of the node.
+   * return the value 'bestGain'.  If a split is made, then splitInfo
+   * and aux may be modified.
    *
-   * This overload is used only for classification.
+   * This overload is used only for classification tasks.
+   *
+   * @code
+   * @article{10.1007/s10994-006-6226-1,
+   *   author = {Geurts, Pierre and Ernst, Damien and Wehenkel, Louis},
+   *   title = {Extremely Randomized Trees},
+   *   year = {2006},
+   *   issue_date = {April 2006},
+   *   publisher = {Kluwer Academic Publishers},
+   *   address = {USA},
+   *   volume = {63},
+   *   number = {1},
+   *   issn = {0885-6125},
+   *   url = {https://doi.org/10.1007/s10994-006-6226-1},
+   *   doi = {10.1007/s10994-006-6226-1},
+   *   journal = {Mach. Learn.},
+   *   month = apr,
+   *   pages = {3–42},
+   *   numpages = {40},
+   * }
+   * @endcode
    *
    * @param bestGain Best gain seen so far (we'll only split if we find gain
    *      better than this).
    * @param data The dimension of data points to check for a split in.
-   * @param numCategories Number of categories in the categorical data.
    * @param labels Labels for each point.
    * @param numClasses Number of classes in the dataset.
    * @param weights Weights associated with labels.
    * @param minimumLeafSize Minimum number of points in a leaf node for
    *      splitting.
+   * @param minimumGainSplit Minimum gain split.
    * @param splitInfo Stores split information on a successful split.
-   * @param minimumGainSplit Minimum  gain split.
    * @param aux Auxiliary split information, which may be modified on a
    *      successful split.
+   * @param splitIfBetterGain When set to true, it will split only when gain is
+   *      better than the current best gain. Otherwise, it always makes a
+   *      split regardless of gain.
    */
-  template<bool UseWeights, typename VecType, typename LabelsType,
-           typename WeightVecType>
+  template<bool UseWeights, typename VecType, typename WeightVecType>
   static double SplitIfBetter(
       const double bestGain,
       const VecType& data,
-      const size_t numCategories,
-      const LabelsType& labels,
+      const arma::Row<size_t>& labels,
       const size_t numClasses,
       const WeightVecType& weights,
       const size_t minimumLeafSize,
       const double minimumGainSplit,
       arma::vec& splitInfo,
-      AuxiliarySplitInfo& aux);
+      AuxiliarySplitInfo& aux,
+      const bool splitIfBetterGain = false);
 
   /**
    * Check if we can split a node.  If we can split a node in a way that
    * improves on 'bestGain', then we return the improved gain.  Otherwise we
-   * return the value 'bestGain'.  If a split is made, then splitInfo and
-   * aux may be modified.  For this particular split type, aux will be empty
-   * and splitInfo will store the number of children of the node.
+   * return the value 'bestGain'.  If a split is made, then splitInfo
+   * and aux may be modified.
    *
-   * This overload is used only for regression.
+   * This overload is used only for regression tasks.
    *
    * @param bestGain Best gain seen so far (we'll only split if we find gain
    *      better than this).
    * @param data The dimension of data points to check for a split in.
-   * @param numCategories Number of categories in the categorical data.
    * @param responses Responses for each point.
    * @param weights Weights associated with responses.
    * @param minimumLeafSize Minimum number of points in a leaf node for
    *      splitting.
+   * @param minimumGainSplit Minimum gain split.
    * @param splitInfo Stores split information on a successful split.
-   * @param minimumGainSplit Minimum  gain split.
    * @param aux Auxiliary split information, which may be modified on a
    *      successful split.
    * @param fitnessFunction The FitnessFunction object instance. It it used to
    *      evaluate the gain for the split.
+   * @param splitIfBetterGain When set to true, it will split only when gain is
+   *      better than the current best gain. Otherwise, it always makes a
+   *      split regardless of gain.
    */
-  template<bool UseWeights, typename VecType, typename ResponsesType,
-           typename WeightVecType>
+  template<bool UseWeights, typename VecType, typename WeightVecType>
   static double SplitIfBetter(
       const double bestGain,
       const VecType& data,
-      const size_t numCategories,
-      const ResponsesType& responses,
+      const arma::rowvec& responses,
       const WeightVecType& weights,
       const size_t minimumLeafSize,
       const double minimumGainSplit,
       double& splitInfo,
       AuxiliarySplitInfo& aux,
-      FitnessFunction& fitnessFunction);
+      FitnessFunction& fitnessFunction,
+      const bool splitIfBetterGain = false);
 
   /**
-   * Return the number of children in the split.
+   * Returns 2, since the binary split always has two children.
    *
    * @param splitInfo Auxiliary information for the split.
-   * @param * (aux) Auxiliary information for the split (Unused).
+   * @param aux Auxiliary split information, which may be modified on a
+   *      successful split.
    */
-  static size_t NumChildren(const double& splitInfo,
-                            const AuxiliarySplitInfo& /* aux */);
+  static size_t NumChildren(const double& /* splitInfo */,
+                            const AuxiliarySplitInfo& /* aux */)
+  {
+    return 2;
+  }
 
   /**
-   * Calculate the direction a point should percolate to.
+   * Given a point, calculate which child it should go to (left or right).
    *
-   * @param point the Point to use.
+   * @param point Point to calculate direction of.
    * @param splitInfo Auxiliary information for the split.
    * @param * (aux) Auxiliary information for the split (Unused).
    */
@@ -133,6 +155,6 @@ class AllCategoricalSplit
 } // namespace mlpack
 
 // Include implementation.
-#include "all_categorical_split_impl.hpp"
+#include "random_binary_numeric_split_impl.hpp"
 
 #endif
